+++ packages/vite-plugin-cloudflare/e2e/basic.test.ts
@@ -1,53 +1,30 @@
 import { describe } from "vitest";
 import { fetchJson, runCommand, test, waitForReady } from "./helpers.js";
 
-const isWindows = process.platform === "win32";
+describe("node compatibility", () => {
+	describe.each(["pnpm --no-store", "npm", "yarn"])("using %s", (pm) => {
+		test("can serve a Worker request", async ({ expect, seed, viteDev }) => {
+			const projectPath = await seed("basic");
+			runCommand(`${pm} install`, projectPath);
 
-const packageManagers = ["pnpm", "npm", "yarn"] as const;
-const commands = ["dev", "preview"] as const;
+			const proc = await viteDev(projectPath);
+			const url = await waitForReady(proc);
+			expect(await fetchJson(url + "/api/")).toEqual({ name: "Cloudflare" });
+		});
+	});
+});
 
-describe("basic e2e tests", () => {
-	describe.each(commands)('with "%s" command', (command) => {
-		// TODO: re-enable `vite preview` tests on Windows (DEVX-1748)
-		describe.skipIf(command === "preview" && isWindows).each(packageManagers)(
-			'with "%s" package manager',
-			(pm) => {
-				describe("node compatibility", () => {
-					test("can serve a Worker request", async ({
-						expect,
-						seed,
-						viteCommand,
-					}) => {
-						const projectPath = await seed("basic");
-						runCommand(`${pm} install`, projectPath);
+// This test checks that wrapped bindings which rely on additional workers with an authed connection to the CF API work
+describe("Workers AI", () => {
+	test("can serve a Worker request", async ({ expect, seed, viteDev }) => {
+		const projectPath = await seed("basic");
+		runCommand(`npm install`, projectPath);
 
-						const proc = await viteCommand(pm, command, projectPath);
-						const url = await waitForReady(proc);
-						expect(await fetchJson(url + "/api/")).toEqual({
-							name: "Cloudflare",
-						});
-					});
-				});
+		const proc = await viteDev(projectPath);
+		const url = await waitForReady(proc);
 
-				// This test checks that wrapped bindings which rely on additional workers with an authed connection to the CF API work
-				describe("Workers AI", () => {
-					test("can serve a Worker request", async ({
-						expect,
-						seed,
-						viteCommand,
-					}) => {
-						const projectPath = await seed("basic");
-						runCommand(`${pm} install`, projectPath);
-
-						const proc = await viteCommand(pm, command, projectPath);
-						const url = await waitForReady(proc);
-
-						expect(await fetchJson(url + "/ai/")).toEqual({
-							response: expect.stringContaining("Workers AI"),
-						});
-					});
-				});
-			}
-		);
+		expect(await fetchJson(url + "/ai/")).toEqual({
+			response: expect.stringContaining("Workers AI"),
+		});
 	});
 });